+++ tests/testthat/test-holiday.R
@@ -190,6 +190,51 @@ test_that("error on incorrect holidays argument", {
   )
 })
 
+test_that("sparse = 'yes' works", {
+  rec <- recipe(~ ., data = test_data)
+
+  dense <- rec %>% 
+    step_holiday(day, sparse = "no", keep_original_cols = FALSE) %>%
+    prep() %>%
+    bake(NULL)
+  dense <- purrr::map(dense, as.integer) %>% tibble::new_tibble()
+  sparse <- rec %>%
+    step_holiday(day, sparse = "yes", keep_original_cols = FALSE) %>%
+    prep() %>%
+    bake(NULL)
+
+  expect_identical(dense, sparse)
+
+  expect_false(any(vapply(dense, sparsevctrs::is_sparse_vector, logical(1))))
+  expect_true(all(vapply(sparse, sparsevctrs::is_sparse_vector, logical(1))))
+})
+
+test_that("sparse argument is backwards compatible", {
+  rec <- recipe(~ ., data = test_data) %>%
+    step_holiday(day) %>%
+    prep()
+
+  exp <- bake(rec, test_data)
+
+  # Simulate old recipe
+  rec$steps[[1]]$sparse <- NULL
+
+  expect_identical(
+    bake(rec, test_data),
+    exp
+  )
+})
+
+test_that(".recipes_toggle_sparse_args works", {
+  rec <- recipe(~ ., data = test_data) %>%
+    step_holiday(day, sparse = "auto", keep_original_cols = TRUE)
+
+  expect_equal(
+    .recipes_estimate_sparsity(rec),
+    364/365
+  )
+})
+
 # Infrastructure ---------------------------------------------------------------
 
 test_that("bake method errors when needed non-standard role columns are missing", {